// backend/src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type DocumentFile = {
  __typename?: 'DocumentFile';
  content: Scalars['String'];
  fileExtension: Scalars['String'];
  id: Scalars['String'];
  listIndex: Scalars['Int'];
  rowsCount: Scalars['Int'];
  title: Scalars['String'];
};

export type Mutation = {
  __typename?: 'Mutation';
  createDocument?: Maybe<DocumentFile>;
  updateDocument?: Maybe<DocumentFile>;
};


export type MutationCreateDocumentArgs = {
  content?: InputMaybe<Scalars['String']>;
  fileExtension?: InputMaybe<Scalars['String']>;
  listIndex?: InputMaybe<Scalars['Int']>;
  rowsCount?: InputMaybe<Scalars['Int']>;
  title?: InputMaybe<Scalars['String']>;
};


export type MutationUpdateDocumentArgs = {
  content?: InputMaybe<Scalars['String']>;
  fileExtension?: InputMaybe<Scalars['String']>;
  id?: InputMaybe<Scalars['String']>;
  listIndex?: InputMaybe<Scalars['Int']>;
  rowsCount?: InputMaybe<Scalars['Int']>;
  title?: InputMaybe<Scalars['String']>;
};

export type Query = {
  __typename?: 'Query';
  document?: Maybe<DocumentFile>;
  documents: Array<DocumentFile>;
};


export type QueryDocumentArgs = {
  id?: InputMaybe<Scalars['String']>;
};



export type ResolverTypeWrapper<T> = Promise<T> | T;


export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
  DocumentFile: ResolverTypeWrapper<DocumentFile>;
  Int: ResolverTypeWrapper<Scalars['Int']>;
  Mutation: ResolverTypeWrapper<{}>;
  Query: ResolverTypeWrapper<{}>;
  String: ResolverTypeWrapper<Scalars['String']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Boolean: Scalars['Boolean'];
  DocumentFile: DocumentFile;
  Int: Scalars['Int'];
  Mutation: {};
  Query: {};
  String: Scalars['String'];
};

export type DocumentFileResolvers<ContextType = any, ParentType extends ResolversParentTypes['DocumentFile'] = ResolversParentTypes['DocumentFile']> = {
  content?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  fileExtension?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  id?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  listIndex?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  rowsCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
  title?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  createDocument?: Resolver<Maybe<ResolversTypes['DocumentFile']>, ParentType, ContextType, Partial<MutationCreateDocumentArgs>>;
  updateDocument?: Resolver<Maybe<ResolversTypes['DocumentFile']>, ParentType, ContextType, Partial<MutationUpdateDocumentArgs>>;
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  document?: Resolver<Maybe<ResolversTypes['DocumentFile']>, ParentType, ContextType, Partial<QueryDocumentArgs>>;
  documents?: Resolver<Array<ResolversTypes['DocumentFile']>, ParentType, ContextType>;
};

export type Resolvers<ContextType = any> = {
  DocumentFile?: DocumentFileResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  Query?: QueryResolvers<ContextType>;
};